Redactor_llama3.1_V1.5/requirements.txt
python-dotenv
beautifulsoup4
requests